+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/mediacodec/MediaCodecRendererTest.java
@@ -734,6 +734,11 @@ public class MediaCodecRendererTest {
     }
 
     @Override
+    public void detachOutputSurface() {
+      throw exceptionSupplier.get();
+    }
+
+    @Override
     public void setParameters(Bundle params) {
       throw exceptionSupplier.get();
     }